### 2019-07-16 #1574 @whedon
[PRE REVIEW]: Pannellum: a lightweight web-based panorama viewer
**Submitting author:** @mpetroff (<a href="http://orcid.org/0000-0002-4436-4215">Matthew Petroff</a>)
**Repository:** <a href="https://github.com/mpetroff/pannellum" target ="_blank">https://github.com/mpetroff/pannellum</a>
**Version:** v2.5.2
