### 2019-07-17 #1574 @leouieda
:wave: Hi @ethan-nelson @vsoch would you be interested in reviewing this submission for JOSS?
