### 2019-07-17 ebc31b7
Fix recovery from load failure (fixes #755).
