### 2019-07-18 060ced6
Fix capitalization in paper reference.
