+++ paper/paper.bib
@@ -10,8 +10,8 @@
 
 @inproceedings{Gede2015,
   author = {M\'{a}ty\'{a}s Gede and Zsuzsanna Ungv\'{a}ri and Klaudia Kiss and G\'{a}bor Nagy},
-  title = {Open-source web-based viewer application for TLS surveys in caves},
-  booktitle = "Proceedings of the 1st {ICA} European Symposium on Cartography ({EuroCarto} 2015)",
+  title = {Open-source web-based viewer application for {TLS} surveys in caves},
+  booktitle = "Proceedings of the 1st {ICA} {European} Symposium on Cartography ({EuroCarto} 2015)",
   year = {2015},
   editor = {Georg Gartner and Haosheng Huang},
   pages = {321--328},
