### 2019-07-21 ec68272
Add parameters for hotSpotDebug, compass and autoRotateInactivityDelay
