### 2019-07-21 ebde708
Add parameter for autoRotateStopDelay
