+++ src/standalone/standalone.js
@@ -32,7 +32,7 @@ function parseURLParameters() {
             case 'hfov': case 'pitch': case 'yaw': case 'haov': case 'vaov':
             case 'minHfov': case 'maxHfov': case 'minPitch': case 'maxPitch':
             case 'minYaw': case 'maxYaw': case 'vOffset': case 'autoRotate':
-            case 'autoRotateInactivityDelay':
+            case 'autoRotateInactivityDelay': case 'autoRotateStopDelay':
                 configFromURL[option] = Number(value);
                 break;
             case 'autoLoad': case 'ignoreGPanoXMP': case 'hotSpotDebug':
