### 2019-07-21 bed0f2e
Add parameter for authorURL
