### 2019-07-21 d1a6574
Add parameters for showZoomCtrl and showControls
