### 2019-07-24 4b2edee
Return JPEG instead of PNG when passing returnImage option to render
Its significantly faster to create a jpeg image than a png.
