### 2019-07-24 14c1246
Add scale option to hotspot
When `true`, the hotspot is scaled to match the fish eye effect of the image.

Hotspots then appear to be embedded in the panorama instead of floating on top
