+++ doc/json-config-parameters.md
@@ -316,6 +316,10 @@ If `clickHandlerFunc` is specified, this function is added as an event handler
 for the hot spot's `click` event. The event object and the contents of
 `clickHandlerArgs` are passed to the function as arguments.
 
+#### `scale` (boolean)
+
+When `true`, the hotspot is scaled to match the fish eye effect of the image
+
 ### `hotSpotDebug` (boolean)
 
 When `true`, the mouse pointer's pitch and yaw are logged to the console when
