### 2019-07-24 6d48997
Add optional pitch argument to startAutoRotate
Before starting rotation, the viewer is panned to `pitch`.

If pitch is not specified, origPitch will be used
