### 2019-08-01 #1574 @leouieda
:wave: Hi @fil would you be interested in reviewing this submission for JOSS?
