### 2019-08-01 b53f18c
Fix docstring punctuation.
