### 2019-08-01 64f7c9c
Add optional pitch argument to startAutoRotate (#761)
* Add optional pitch argument to startAutoRotate

Before starting rotation, the viewer is panned to `pitch`.

If pitch is not specified, origPitch will be used
