### 2019-08-02 d773cc9
Fix regression with auto-rotate stopping early (fixes #764).
This regression was introduced in 12516e69b90e8ab4c6f42b64560aef92ae539bd0.
