### 2019-08-02 2a2a4c2
Fix hot spot scale parameter documentation.
