### 2019-08-02 3bbc557
Add scale option to hot spot (#760)
* Add scale option to hot spot

When `true`, the hot spot is scaled to match the fish eye effect of the image.

Hot spots then appear to be embedded in the panorama instead of floating on top.
