### 2019-08-05 #1628 @whedon
[REVIEW]: Pannellum: a lightweight web-based panorama viewer
**Submitting author:** @mpetroff (<a href="http://orcid.org/0000-0002-4436-4215">Matthew Petroff</a>)
**Repository:** <a href="https://github.com/mpetroff/pannellum" target ="_blank">https://github.com/mpetroff/pannellum</a>
**Version:** v2.5.3

- [x] **License:** Does the repository contain a plain-text LICENSE file with the contents of an [OSI approved](https://opensource.org/licenses/alphabetical) software license? **The repository contains a COPYING file, which is detected by GitHub as a license, and follows suit of other GNU software (so I approve)**
