### 2019-08-05 #1628 @leouieda
:wave: @vsoch @Fil thank you for agreeing to review! Please feel free to post your comments on this thread and/or open issues or pull-requests on the software repository. If opening issues or PRs, please remember to mention `openjournals/joss-reviews#1628` so that they are linked back to this thread. This makes it easier for me to monitor the review progress.

@vsoch @Fil @mpetroff please feel free to ask if you have any questions about the review process.
