### 2019-08-06 mpetroff/pannellum#769 @vsoch
JoSS Review #1: Paper Comments
hey @mpetroff ! This is for the review at https://github.com/openjournals/joss-reviews/issues/1628. I'll be opening different issues to coincide with groups of comments and suggestions. This one is about the paper itself. Overall, great job on it!

----------------------------------------------------------

## Paper Comments

The introduction flows beautifully - opening with a description of the tool,
and just as I had finished reading the first paragraph (and was thinking "What other
tools have existed?") the second paragraph opens to describe the history of the technology.

 1. Could you describe more what an integration with maps would look like? Mousing over a map
and opening a panorama for a point, or (within a panorama) ascribing locations to things?
 2. What is a panoramic video?
 3. It might be interesting to tell the reader how the term "Pannellum" originated - what does it mean?

The paper is succinct and good. The one recommendation I'd suggest is to add (somewhere) in the text a link for the reader to easily jump to documentation, examples, etc. The JoSS Template automatically provides a link to the repository, but we would want to go directly to the documentation, or another reference in the paper (that isn't in the references, if appropriate).
