### 2019-08-06 mpetroff/pannellum#770 @vsoch
JoSS Review #2: Repository Comments
This issue is for (superficial) comments about files / structure of the repository, again for https://github.com/openjournals/joss-reviews/issues/1628. Everything required is included, and the points I'd like to confirm are:

 1. if a CONTRIBUTING.md should be added in the .github folder to be programatically discoverable
 2. Choice for naming the LICENSE file "COPYING" ?

That's it. :) I still haven't reviewed the software itself - doing that now!

## Repository Comments

There isn't a CONTRIBUTING.md, but the author provides text in the README. In that
 programmatic discovery can be done by including a CONTRIBUTING.md, I'd suggest to the author to consider adding one (with the content from the README) to the .github folder.

The LICENSE file isn't kept with the repository, but mentioned (and linked) with the README. It looks like the file is named COPYING. GitHub seems to detect it, so if this is the author's preference, it's okay to do. Most folks will look for a LICENSE, if that makes a difference.

I can confirm that @mpetroff has made the greatest contributions, and all others (total of 31) look like tiny bug fixes, etc. It's most appropriate that only he is the author.
