### 2019-08-06 mpetroff/pannellum#771 @vsoch
JoSS Review #3: Getting Started
This is for https://github.com/openjournals/joss-reviews/issues/1628

As a user, I'm expecting to hit the README and be walked through a "getting started" set of instructions. This is here to some extent, but it's not made explicitly clear - you have beautiful documentation pages with examples, and tools to quickly generate panoramas, and the reader should know about them immediately! I would suggest adding a few lines that explicitly give the user choices, something like:

### Getting Started

 1. **Examples**: If you want to see the tool in action, we have a set of [examples](https://pannellum.org/documentation/examples/simple-example/) ready to go that you can view.
 2. **Test Your Own**: If you want to quickly test Pannellum with a panorama, try out our [quick tutorial](https://pannellum.org/documentation/overview/tutorial/).
 3. **Build Your Own**: To prepare your own server with Pannellum, follow the [instructions here](https://github.com/mpetroff/pannellum#how-to-use).

Right at the start of the README, after the About section.
