### 2019-08-06 mpetroff/pannellum#771 @vsoch
Also - for the third (Build your own) you mention a build/pannellum.htm - but glancing over the top level repository structure I don't see that. I think you would want to more explicitly walk the user through these steps, e.g.,:

```bash
git clone https://github.com/mpetroff/pannellum
cd pannellum/<some folder>

# steps for generating

# steps for previewing with python http.server or similar
```
