### 2019-08-06 mpetroff/pannellum#771 @vsoch
Following the instructions to use generate.py, I do the following:

```bash
 git clone https://github.com/mpetroff/pannellum
cd pannellum/

# I downloaded this from Google Photos, you might provide a wget command for the user to get an example
mv panorama.jpg utils/multires/
cd utils/multires/
python generate.py panorama.jpg
```
This is the command provided in the README, but I'm being told that an argument is missing (that isn't documented in the help):

```bash
$ python generate.py panorama.jpg
usage: generate.py [-h] [-C] [-H HAOV] [-F HFOV] [-V VAOV] [-O VOFFSET]
                   [-e HORIZON] [-o OUTPUT] [-s TILESIZE] [-f FALLBACKSIZE]
                   [-c CUBESIZE] [-b BACKGROUNDCOLOR] [-B] [-a] [-q QUALITY]
                   [--png] -n EXECUTABLE [-G] [-d]
                   INPUT
generate.py: error: the following arguments are required: -n/--nona
```
