### 2019-08-06 mpetroff/pannellum#771 @vsoch
I'll test out this "How to use" section when you've had a chance to fix it up a bit! :)
