### 2019-08-06 mpetroff/pannellum#772 @vsoch
JoSS Review #4: Examples Folder
This is for https://github.com/openjournals/joss-reviews/issues/1628

Are the examples intended to work with a local web server? I can cd into the examples folder, start the server:

```bash
$ python -m http.server 9999
```
And then go to the page, but I'm getting a 404:
![image](https://user-images.githubusercontent.com/814322/62551162-7c536e80-b839-11e9-9d04-9fdcb0cea6ed.png)

Is port 80 required, so a user wouldn't be allowed to test with python, etc.?
