### 2019-08-06 mpetroff/pannellum#773 @vsoch
JoSS Review #5: Community Guidelines
This is for https://github.com/openjournals/joss-reviews/issues/1628

I mentioned possibly adding a CONTRIBUTING.md in #770 , and I'm looking over this checkpoint:

> - [ ] **Community guidelines:** Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

And I'd like to suggest adding a "Support" section to the README, where you state (and link) to your preferred method for users to ask questions, report bugs, etc. it's implied to be GitHub issues, but it's helpful to state it explicitly.
