### 2019-08-06 mpetroff/pannellum#774 @vsoch
JoSS Review #6: Automated Testing
This is for https://github.com/openjournals/joss-reviews/issues/1628

I'd like to see automated testing, to ensure that tests are run with every pull request and change (and for the user not to need to find / download a geckodriver to run them). CircleCI and Travis can handle headless testing, so you could check those out (or others that you have preference to use!)  The recipe should handle installing the necessary python packages, obtaining (or keeping a geckodriver with the code) and then ensuring that they pass.
