### 2019-08-07 fb918f1
Update JOSS paper to address comments by @vsoch (issue #769).
