+++ paper/paper.bib
@@ -98,3 +98,12 @@
   type = {{W3C} Recommendation},
   institution = {World Wide Web Consortium ({W3C})},
 }
+
+@online{OSM2018,
+  author = {{OpenStreetMap}},
+  title = {{Bing} {Streetside} imagery now available in {OpenStreetMap} {iD} editor},
+  year = {2018},
+  month = jul,
+  url = {https://blog.openstreetmap.org/2018/07/03/bing-streetside-in-id/},
+  urldate = {2019-08-06},
+}
