### 2019-08-07 mpetroff/pannellum#769 @mpetroff
Thank you for the detailed comments and quick response! I'll try to work through them this week.

>     1. Could you describe more what an integration with maps would look like? Mousing over a map
>        and opening a panorama for a point, or (within a panorama) ascribing locations to things?

I added a description and also added an extra reference, to a [blog post with an example](https://blog.openstreetmap.org/2018/07/03/bing-streetside-in-id/).

>     2. What is a panoramic video?

I added an explanation.

>     3. It might be interesting to tell the reader how the term "Pannellum" originated - what does it mean?

I had originally included this but decided to remove it as superfluous before submitting the manuscript; I added it back in.

> The paper is succinct and good. The one recommendation I'd suggest is to add (somewhere) in the text a link for the reader to easily jump to documentation, examples, etc. The JoSS Template automatically provides a link to the repository, but we would want to go directly to the documentation, or another reference in the paper (that isn't in the references, if appropriate).

I added a few links at the end of the paper.

The changes were made in fb918f1f36ced9c7f8c60d82ca27ad40a2074dc5: https://github.com/openjournals/joss-papers/blob/joss.01628/joss.01628/10.21105.joss.01628.pdf
