### 2019-08-07 5e682d0
Add `CONTRIBUTING.md` file per JOSS review suggestion (#770).
