### 2019-08-07 mpetroff/pannellum#770 @mpetroff
>     1. if a CONTRIBUTING.md should be added in the .github folder to be programatically discoverable

I added this file in 5e682d0ff346687a7d98437e5d0b176c1b173075.

>     2. Choice for naming the LICENSE file "COPYING" ?

This is a fairly common / traditional naming choice used by most GNU projects, e.g., [GCC](https://gcc.gnu.org/git/?p=gcc.git;a=blob;f=COPYING;h=623b6258a134210f0b0ada106fdaab7f0370d9c5;hb=HEAD) or [Coreutils](https://github.com/coreutils/coreutils/blob/master/COPYING), as well as other common software such as [Git](https://github.com/git/git/blob/master/COPYING). It's also used by the first open source project I became involved with, [Hugin](https://sourceforge.net/p/hugin/hugin/ci/default/tree/COPYING.txt), which is the primary reason for the naming choice.
