### 2019-08-07 982242c
Clarify directions for viewing examples (#772).
