### 2019-08-07 mpetroff/pannellum#772 @mpetroff
I added some clarifying instructions in 982242c7f9c5060fb02b684c25bd39ba02f797b9 for how to view the examples.

You encountered an error loading an `<iframe>` of a file in the `src/standalone` directory, since the web server was started from the `examples` directory, not the repository root. Port 80 is not required.
