### 2019-08-07 405bdc7
Explicitly state prefered method for seeking support (#773).
