### 2019-08-07 mpetroff/pannellum#773 @mpetroff
I added a `CONTRIBUTING.md` file in 5e682d0ff346687a7d98437e5d0b176c1b173075 and added a "Seeking support" section to the readme in 405bdc717de33da52e3c3a7b2ea5c37cf1c35fb7.
