### 2019-08-07 mpetroff/pannellum#774 @mpetroff
WebGL unfortunately complicates headless testing, and I wasn't able to get it to work properly with Travis when I previously tried. I ended up setting it up using [Travis](https://travis-ci.org/mpetroff/pannellum) + Sauce Labs, which, unfortunately, doesn't work for pull requests due to security concerns, i.e., stealing the SauceLabs private API key, but does run for commits to `master`. I'll take another look at trying to get the tests to run directly on Travis.
