### 2019-08-07 2d0230d
Another attempt at headless testing with Travis CI.
Try to replace Sauce Labs, so CI works for pull requests (#774).

Based on configuration at https://github.com/stackgl/headless-gl/blob/60c73499fc96b8e61da1058d2ed6a80411890b53/README.md.
