+++ .travis.yml
@@ -1,13 +1,19 @@
 language: generic
 dist: xenial
+services:
+  - xvfb
 addons:
-  sauce_connect: true
+  firefox: latest
 before_install:
-  - sudo apt-get install -y python3-pillow python3-numpy python3-pip
+  - sudo apt-get install -y python3-pillow python3-numpy python3-pip mesa-utils libgl1-mesa-dri libglapi-mesa libosmesa6
   - sudo pip3 install selenium
+install:
+  - wget https://github.com/mozilla/geckodriver/releases/download/v0.24.0/geckodriver-v0.24.0-linux64.tar.gz
+  - tar -xf geckodriver-v0.24.0-linux64.tar.gz
+  - sudo mv geckodriver /usr/local/bin
 jobs:
   include:
     - stage: build
       script: python3 utils/build/build.py
     - stage: test
-      script: python3 tests/run_tests.py
+      script: xvfb-run -a python3 tests/run_tests.py
