### 2019-08-07 9e23ee0
Print browser console logs when tests fail.
