### 2019-08-07 46c89c4
Revert "Another attempt at headless testing with Travis CI."
This reverts commits 2d0230dd4e1ae77a1d7dd8705fedf0f5f4c40156,
9e23ee0b5708bb96552d15571ec966f604a43301,
e996faa86c9dcc4c3e4a859cbb75567d4e3ace91,
and d204ac06d5127135258584d886b089f7092c5856.
