### 2019-08-07 mpetroff/pannellum#774 @mpetroff
Still no luck getting it to work. I'm out of ideas for getting it to work with Firefox. I guess I could try getting it to work with Chromium.
