### 2019-08-07 6415759
Stop animate loop running after destroy has been called
