### 2019-08-07 mpetroff/pannellum#769 @vsoch
- The added example with Bing streetmaps (and your description) much better illustrates what an integration entails.
 - I'm glad that you added the explanation of the name back in - it's something that has chance to be forgotten, and it's interesting, and this will preserve the history of the tool.
 - The addition of the links to examples at the end is spot on what I had in mind:

![image](https://user-images.githubusercontent.com/814322/62623463-e2e69400-b8ee-11e9-8920-b3e073962bea.png)

The typical reader would finish reading, and then want to go see / experiment with the tool. This will take him or her right to it!
