### 2019-08-07 mpetroff/pannellum#769 @vsoch
You've addressed all my points here, closing issue.
