### 2019-08-07 mpetroff/pannellum#770 @vsoch
Awesome - we have a CONTRIBUTING.md! It seems trivial, but (having done analysis of repositories for fun myself) it's really good to have it - people know to go looking for it.

For the LICENSE, this will be ok to leave as COPYING, since it's a standard that others follow (and I didn't know about!) and GitHub seems to pick it up. My concerns have been well addressed, thanks for the details @mpetroff.
