### 2019-08-07 mpetroff/pannellum#772 @vsoch
We're in business! The examples instructions (for staying in the root) are easy to miss because you don't directly give the user something to copy paste. For example, you could do something like:

## Examples

Examples using both the minified version and the version in the src directory are included in the examples directory. These can be viewed by starting a local web server in the root of the repository:

```bash
python3 -m http.server
```
in the directory containing this readme file, and then navigating to the hosted HTML files using a web browser; note that the examples use files from the src directory, so **the web server must be started from the repository root, not the examples directory**.

-------------------------------------------------------

The above is just a tiny difference - but it gives the user the command off the bat, and then bolds the point about where they should be.

I tested all the examples - and for the most part, we are in business! There is one that still doesn't seem to work:

![image](https://user-images.githubusercontent.com/814322/62624014-255ca080-b8f0-11e9-8da9-0251c360ae8e.png)
