### 2019-08-07 mpetroff/pannellum#773 @vsoch
Spot on! Closing issue.
