### 2019-08-07 13ceeb6
Squashed commits for fixing CI.
try adding circle config (#1)

* adding automated tests to run on Circle with selenium, still need to add screenshots
Signed-off-by: Vanessa Sochat <vsochat@stanford.edu>

Try taking screenshot (#2)

Adding screenshots (working) back in to tests, ready to PR to upstream
Signed-off-by: Vanessa Sochat <vsochat@stanford.edu>

Don't include driver binaries.

Take screenshot of only viewer element.

Don't remove Travis config, at least for now.

Run Black on tests script.

Fix automated tests for Firefox and some formatting changes.

Update test documentation.
