+++ .circleci/config.yml
@@ -1,0 +1,27 @@
+# Python CircleCI 2.0 configuration file
+# Check https://circleci.com/docs/2.0/language-python/ for more details
+#
+version: 2
+jobs:
+  build:
+    docker:
+      - image: selenium/standalone-chrome:3.141
+
+    working_directory: ~/repo
+
+    steps:
+      - checkout
+
+      - run:
+          name: install dependencies
+          command: |
+            sudo apt-get update && \
+            sudo apt-get install -y python3 python3-dev default-jdk python3-pillow python3-numpy python3-pip && \
+            sudo apt-get install -y libjpeg-dev libfreetype6 libfreetype6-dev zlib1g-dev optipng
+            sudo pip3 install selenium Pillow
+
+      - run:
+          name: run tests
+          command: |
+            python3 utils/build/build.py
+            python3 tests/run_tests.py
