### 2019-08-07 9bb7e4b
try adding circle config (#1)
* adding automated tests to run on Circle with selenium, still need to add screenshots
Signed-off-by: Vanessa Sochat <vsochat@stanford.edu>
