+++ tests/run_tests.py
@@ -1,141 +1,239 @@
-"""
-Selenium-based test suite for Pannellum
+#!/usr/bin/env python
 
-Dependencies:
-Python 3, Selenium Python bindings, Firefox, geckodriver, Pillow, NumPy
-"""
+'''
 
-import http.server
-import time
-import threading
-import io
-import subprocess
-import os
-import numpy as np
-from PIL import Image, ImageChops
+Run tests for Pannellum, set up with Continuous Integration.
+Contributed by Vanessa Sochat, JoSS Review 2019.
+See the project repository for licensing information.
+
+'''
+
+from selenium.common.exceptions import TimeoutException
+from selenium.webdriver.chrome.options import Options
+from selenium.webdriver.common.keys import Keys
+from random import choice
+from threading import Thread
 from selenium import webdriver
+from http.server import SimpleHTTPRequestHandler
+from socketserver import TCPServer
+from PIL import Image, ImageChops
+import argparse
+import json
+import io
+import os
+import re
+import shutil
+import subprocess
+import sys
+import time
+import webbrowser
 
 
-# Set to true to create a new set of reference images
-CREATE_REF = False
+class PannellumServer(SimpleHTTPRequestHandler):
+    '''here we subclass SimpleHTTPServer to capture error messages
+    '''
+    def log_message(self, format, *args):
+        '''log to standard error with a date time string,
+            and then call any subclass specific logging functions
+        '''
+        sys.stderr.write("%s - - [%s] %s\n" %
+                     (self.address_string(),
+                      self.log_date_time_string(),
+                      format%args))
+
+        # Workaround for error trying to GET html
+        if not re.search("div", format%args) and not re.search("function", format%args):
+            if re.search("404", format%args):
+                raise IOError(format%args)
+
+    def log_error(self, format, *args):
+        '''catch errors in the log_messages instead
+        '''
+        pass
 
 
-# Run web server
-print("Starting web server...")
-os.chdir(os.path.dirname(os.path.abspath(__file__)))  # cd to script dir
-os.chdir("..")
-httpd = http.server.HTTPServer(
-    ("localhost", 8000), http.server.SimpleHTTPRequestHandler
-)
-thread = threading.Thread(None, httpd.serve_forever)
-thread.start()
+class PannellumTester(object):
+    ''' bring up a server with a testing robot        
+    '''
+  
+    def __init__(self, **kwargs):
+        self.Handler = PannellumServer
+        if "port" in kwargs:
+            self.port = kwargs['port']
+        else:
+            self.port = choice(range(8000, 9999))
+        print('Selected port is %s' % self.port)
+        self.httpd = TCPServer(("", self.port), self.Handler)
+        self.server = Thread(target=self.httpd.serve_forever)
+        self.server.setDaemon(True)
+        self.server.start()
+        self.started = True
+        self.pause_time = 100
+        self.browser = None
+        self.headless = False
+        self.display = None
+        self.driver = "Chrome"
+        if "browser" in kwargs:
+            self.driver = kwargs['browser']
 
 
-# Create a new instance of the Firefox driver
-print("Starting web driver...")
-if os.environ.get("TRAVIS_JOB_NUMBER"):
-    # Configuration for Travis CI / Sauce Labs testing
-    driver = webdriver.Remote(
-        command_executor="https://ondemand.saucelabs.com:443/wd/hub",
-        desired_capabilities={
-            "username": os.environ["SAUCE_USERNAME"],
-            "accessKey": os.environ["SAUCE_ACCESS_KEY"],
-            "tunnel-identifier": os.environ["TRAVIS_JOB_NUMBER"],
-            "build": os.environ["TRAVIS_JOB_NUMBER"],
[…]
