### 2019-08-07 mpetroff/pannellum#774 @vsoch
hey @mpetroff - we have to keep trying! To help, I created a CircleCI configuration with tests using Chrome to run selenium headless. The passing test (can you see?) https://circleci.com/gh/vsoch/pannellum/15 and then the added configuration is [here](https://github.com/vsoch/pannellum/blob/master/.circleci/config.yml), and the testing script (with an argument parser) is [here](https://github.com/vsoch/pannellum/blob/master/tests/run_tests.py). The reason I haven't pull requested yet is because the screenshots weren't working for me, just locally, so I removed them to get basic tests working and then also working in the CI. If we can debug the screenshots, then you should have fully automated tests. :)

The error I get locally is:

```
    raise exception_class(value)
selenium.common.exceptions.WebDriverException: Message: unknown command: session/561f7791117f8823a289d498d9ffcf72/element/0.2280353649737068-1/screenshot
```
And I'm about to test on CircleCI to see if it happens there too. Have you seen anything like this?
