### 2019-08-07 460ccec
Try taking screenshot (#2)
Adding screenshots (working) back in to tests, ready to PR to upstream
Signed-off-by: Vanessa Sochat <vsochat@stanford.edu>
