### 2019-08-07 mpetroff/pannellum#774 @vsoch
Got it all working! See https://github.com/mpetroff/pannellum/pull/777
