### 2019-08-07 mpetroff/pannellum#771 @Fil
Same for me & I would add that the package.json file mentions "main": "build/pannellum.js", which is not there and doesn't get built automatically (at least I couldn't find how).
