### 2019-08-07 mpetroff/pannellum#771 @vsoch
@Fil it's hiding in the utils folder, and I think @mpetroff is going to clean this up a bit!
