### 2019-08-07 mpetroff/pannellum#771 @Fil
True :) and I found another copy in src/ which doesn't seem to need a build step.

Other than that, my review is complete and as soon as the installation / dependencies are cleaned up I'll check off the last items. 👍

Very nice seeing pannellum featured in JOSS!
