### 2019-08-08 ac72b8c
Update examples documentation per #772.
