### 2019-08-08 e622878
Remove CircleCI configuration, since Travis CI is working now.
