### 2019-08-08 mpetroff/pannellum#772 @mpetroff
I added this in ac72b8cabf90685d79e26f187d51190a12d3376e. Thanks for the helpful suggestion.
