### 2019-08-08 mpetroff/pannellum#774 @mpetroff
#777 was applied in a squashed commit in 13ceeb605601776a57c69f1ba00be70e5d968735 and applied to Travis in 6675d15d9db90b80fa91426ea636a80d3e56d42e. Continuous integration should now be working for both `master` and pull requests.

More discussion in https://github.com/mpetroff/pannellum/pull/777#issuecomment-519396603.
