### 2019-08-08 mpetroff/pannellum#771 @mpetroff
I'll see about addressing this tomorrow.
