### 2019-08-08 mpetroff/pannellum#774 @vsoch
Great! And here is the green testing, for future reference. https://travis-ci.org/mpetroff/pannellum I'm glad this is working, and glad to help :)
