### 2019-08-08 mpetroff/pannellum#772 @vsoch
okay, minified version is working after build!

![image](https://user-images.githubusercontent.com/814322/62704168-c1ea7580-b9b8-11e9-941b-45f0ffabcf85.png)

Is that Jordan Pond at Acadia National Park? I grew up in rural NH so we went there a few times, really beautiful :)
