### 2019-08-08 mpetroff/pannellum#772 @mpetroff
Yes, it is, circa 2009.
