### 2019-08-09 539fdc2
Substantially rework parts of readme to try to address #771.
