### 2019-08-09 mpetroff/pannellum#771 @mpetroff
I made some significant changes to the readme in 539fdc27493bd70fe5e9c8426c4e8b7b559a4e4e to try to address the concerns raised here. Addressing specific issues:

For the issue with the `generate.py` script:
The issue is due to `nona` (from [Hugin](http://hugin.sourceforge.net/)) not being found on the `$PATH`. When it isn't found, the `-n/--nona` flag is required to manually specify its location. This was previously documented in the readme, but I clarified it a bit more now. It was also documented in the help (`./generate.py --help`), so you must have missed it; to help with this, I added a more explicit error message for the missing parameter in f9749f9b5e4e4e421b52783c2f3b9297592fbb40.

For `build/pannellum.js`:
I added more explicit mentions to the files in the `build` directory and how they're created with the readme changes.
