### 2019-08-09 mpetroff/pannellum#771 @Fil
Thanks, it clarifies what I was missing.
