### 2019-08-09 a886ab3
adding complete example with Dockerfile to use generate.py
Signed-off-by: Vanessa Sochat <vsochat@stanford.edu>
