### 2019-08-09 mpetroff/pannellum#771 @vsoch
The getting started section is spot on! It provides the missing transition between reading the about and knowing what to do first.

For record, here is what was printed before updates to the client - if I run the generate.py it just tells me an argument is missing:
```
$ python3 generate.py panorama.jpg
usage: generate.py [-h] [-C] [-H HAOV] [-F HFOV] [-V VAOV] [-O VOFFSET]
                   [-e HORIZON] [-o OUTPUT] [-s TILESIZE] [-f FALLBACKSIZE]
                   [-c CUBESIZE] [-b BACKGROUNDCOLOR] [-B] [-a] [-q QUALITY]
                   [--png] -n EXECUTABLE [-G] [-d]
                   INPUT
generate.py: error: the following arguments are required: -n/--nona
```
and then when I ask for help, it looks like nona was there, but it's after -n EXECUTABLE so I just missed it :P User error!

```
$ python3 generate.py --help| grep nona
  -n EXECUTABLE, --nona EXECUTABLE
                        location of the nona executable to use (default: None)
  -G, --gpu             perform image remapping by nona on the GPU (default:
```
Anyway let's test the fixed version now!

```
$ python3 generate.py panorama.jpg
IMPORTANT: The location of nona must be specified with -n, since it was not
           found on the PATH!

usage: generate.py [-h] [-C] [-H HAOV] [-F HFOV] [-V VAOV] [-O VOFFSET]
                   [-e HORIZON] [-o OUTPUT] [-s TILESIZE] [-f FALLBACKSIZE]
                   [-c CUBESIZE] [-b BACKGROUNDCOLOR] [-B] [-a] [-q QUALITY]
                   [--png] -n EXECUTABLE [-G] [-d]
                   INPUT
generate.py: error: the following arguments are required: -n/--nona
```
Even though I don't know what nona is, the error message is loud and clear, and I can Google search, etc. to figure it out.

Okay just finished testing! What I wound up doing is hitting dependency issues on my host, and I built a container that runs it easily without a hitch! What I'm going to do is open a PR to add a simple Dockerfile to the generate folder, one the user can easily build and run the script without running into issue. Stay tuned!
