### 2019-08-09 mpetroff/pannellum#771 @vsoch
okay see https://github.com/mpetroff/pannellum/pull/779!
