### 2019-08-10 1f06b1e
Rewrite multires readme to not just be about Docker.
