### 2019-08-10 37329ec
Simplify dependencies in Dockerfile.
