### 2019-08-10 mpetroff/pannellum#771 @mpetroff
#779, which provided a Dockerfile, has now been merged.
