### 2019-08-10 4bb9df8
adding missing dockerfile dependency
Signed-off-by: Vanessa Sochat <vsochat@stanford.edu>
