+++ utils/multires/Dockerfile
@@ -5,7 +5,7 @@ FROM ubuntu:18.04
 
 ENV DEBIAN_FRONTEND noninteractive
 RUN apt-get update && \
-    apt-get install -y python3 python3-pil hugin-tools
+    apt-get install -y python3 python3-dev python3-pil hugin-tools
 
 ADD generate.py /generate.py
 ENTRYPOINT ["python3", "/generate.py"]
