### 2019-08-10 mpetroff/pannellum#771 @vsoch
Looks great now! I've gone through each of the use cases in the getting started, and it's just what a new user would want for a gentle introduction. Closing issue!
