### 2019-08-10 #1628 @vsoch
My review is complete! @mpetroff has addressed all of my issues (linked above), and specifically:
 - the paper is well written, clearly communicates a need, walks the reader through some background and how it works, and finishes by directing to the documentation
 - the repository now has a clean set of examples, one of which is driven by a Dockerfile to lower the barrier to entry
 - testing is automated
 - community guidelines are clear

Great work on this @mpetroff! I definitely approve this software for publication in JoSS.
