### 2019-08-11 484d4c5
fixed that in tiled panoramas some tiles at the bottom or top were not refined
