### 2019-08-11 5a058fa
Minor proofreading changes to JOSS paper.
Garner's Modern English Usage (4th ed.) recommends "webpage" over "web page."
