### 2019-08-16 6fe5193
Fix behavior of `multiResMinHfov` to match docs (fixes #783).
