### 2019-08-21 e1d4c9a
Improved recovery from failed scene loading (#755).
