+++ changelog.md
@@ -2,6 +2,30 @@ Changelog
 =========
 
 
+Changes in Pannellum 2.5.3 (2019-08-21)
+---------------------------------------
+
+Bugfixes:
+ - Fixed loading scenes after previous load failure
+ - Fixed regression that caused auto-rotation to stop after one revolution
+ - Fixed behavior of `multiResMinHfov` to match documentation;
+   default multi-resolution `minHfov` behavior now matches pre-v2.5
+
+Improvements:
+ - Added optional `scale` parameter to scale hot spots while zooming
+ - Improved recovery from failed scene loading
+
+API improvements:
+ - Added optional pitch argument to `startAutoRotate`
+
+Other:
+ - Added Journal of Open Source Software (JOSS) manuscript
+ - Numerous documentation improvements per JOSS review (thanks @vsoch and @Fil)
+ - Improved continuous integration and automated testing support
+ - Clarified `generate.py` error messages
+ - Added Dockerfile for `generate.py`
+
+
 Changes in Pannellum 2.5.2 (2019-07-13)
 ---------------------------------------
 
