### 2019-08-21 df4c45d
Move `load` event to fire after rendering instead of before (#768).
