+++ src/js/pannellum.js
@@ -1678,10 +1678,10 @@ function renderInitCallback() {
         preview = undefined;
     }
     loaded = true;
-
-    fireEvent('load');
     
     animateInit();
+
+    fireEvent('load');
 }
 
 /**
