### 2019-08-21 #1628 @mpetroff
@leouieda The review was completed a week and a half ago. I just released v2.5.3 with the changes made as a result of the review (mostly documentation / testing improvements) and a few unrelated bug fixes.
