### 2019-08-30 8b9cc48
Request device orientation permission (untested).
This is needed for iOS 13+ for device orientation mode to work.
