### 2019-08-31 #1628 @leouieda
@vsoch @Fil thank you for your reviews :confetti_ball:

@mpetroff I had a final look through the paper and since both reviewers approve I'm happy to move forward with acceptance of your paper :tada: Here are the steps you need to take now:

- Make a release with the latest changes from the review and post the version number here.
- Archive the release on [Zenodo](https://zenodo.org/) and post the DOI here (make sure the title and author list matches the JOSS paper, if possible).

Let me know if you need help with any of these steps.
