### 2019-08-31 #1628 @mpetroff
> * Double check authors and affiliations (including ORCIDs)

Done.

> * Make a release with the latest changes from the review and post the version number here.

v2.5.3

> * Archive the release on [Zenodo](https://zenodo.org/) and post the DOI here (make sure the title and author list matches the JOSS paper, if possible).

[10.5281/zenodo.3373725](https://doi.org/10.5281/zenodo.3373725)
