### 2019-08-31 #1628 @mpetroff
The concept DOI is [10.5281/zenodo.3334433](https://doi.org/10.5281/zenodo.3334433); I don't know which is preferred for JOSS.
