### 2019-08-31 #1628 @leouieda
@openjournals/joss-eics this submission is ready for publication :tada:
