### 2019-08-31 #1628 @labarba
@mpetroff — Could you edit the metadata of the Zenodo archive so the title matches the JOSS paper?
