### 2019-08-31 #1628 @labarba
Also, you've got three new commits since v2.5.3. Do you need to up the patch version for the accepted software?
