### 2019-08-31 #1628 @mpetroff
I updated the Zenodo title. The three most recent commits are unrelated to the review and are intended for v2.6.0.
