### 2019-08-31 #1628 @vsoch
And @whedon is channeling Hal... is the predicted future from the movie coming true? Things just got real! :scream:

Happy Saturday everyone :)
