### 2019-08-31 #1628 @labarba
It worked. Congratulations, @mpetroff, your JOSS paper is published! 🚀

Big thanks to our editor: @leouieda, and the reviewers: @vsoch, @Fil — thank you for your contributions to JOSS 🙏
